requirements.txt
fastapi
uvicorn
openai
beautifulsoup4
requests
faiss-cpu
python-dotenv
sentence-transformers
google-generativeai==0.3.2
